Move Endpoint enum from "util" to "io.http"
