+++ src/main/java/io/github/hypixel_api_wrapper/io/http/Endpoint.java
@@ -1,4 +1,4 @@
-package io.github.hypixel_api_wrapper.util;
+package io.github.hypixel_api_wrapper.io.http;
 
 import java.net.MalformedURLException;
 import java.net.URL;
