+++ src/main/java/io/github/hypixel_api_wrapper/io/http/RequestFactory.java
@@ -1,7 +1,6 @@
 package io.github.hypixel_api_wrapper.io.http;
 
 import io.github.hypixel_api_wrapper.io.caching.CachingStrategy;
-import io.github.hypixel_api_wrapper.util.Endpoint;
 import java.io.IOException;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
